BasicPlayerController
Made base functionality for Player: translation with WASD, rotation with mouse, ability to sprint, jump, and shoot (shoot is currently just a raycast that grabs the name of the object if an object is hit and has a collider). Also found some character models on Mixamo that may be of interest to the group.